IPROTO-85 Refactor ProtoTypeMetadata for better representation of imported types
* extract ImportedProtoTypeMetadata class
